ajouter autodoc, corriger docstring avec doctest
