+++ Makefile
@@ -1,19 +1,20 @@
-all: doc lint test
+# Minimal makefile for Sphinx documentation
+#
 
-doc:
-	make -C docs html
+# You can set these variables from the command line.
+SPHINXOPTS    =
+SPHINXBUILD   = sphinx-build
+SPHINXPROJ    = gitkv
+SOURCEDIR     = .
+BUILDDIR      = _build
 
-test:
-	coverage run -m doctest ./tests/__init__.py
+# Put it first so that "make" without argument is like "make help".
+help:
+	@$(SPHINXBUILD) -M help "$(SOURCEDIR)" "$(BUILDDIR)" $(SPHINXOPTS) $(O)
 
-lint:
-	flake8 $$(find . -type f -name '*.py')
+.PHONY: help Makefile
 
-fixme:
-	find . -type f | xargs grep --color -H -n -i fixme
-
-todo:
-	find . -type f | xargs grep --color -H -n -i todo
-
-live:
-	find . -type f -name '*.py' | entr make test
+# Catch-all target: route all unknown targets to Sphinx using the new
+# "make mode" option.  $(O) is meant as a shortcut for $(SPHINXOPTS).
+%: Makefile
+	@$(SPHINXBUILD) -M $@ "$(SOURCEDIR)" "$(BUILDDIR)" $(SPHINXOPTS) $(O)
