+++ docs/conf.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 #
 # gitkv documentation build configuration file, created by
-# sphinx-quickstart on Wed Mar 15 11:49:51 2017.
+# sphinx-quickstart on Thu Mar 23 13:28:08 2017.
 #
 # This file is execfile()d with the current directory set to its
 # containing dir.
@@ -31,8 +31,10 @@
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = ['sphinx.ext.doctest',
-    'sphinx.ext.coverage']
+extensions = ['sphinx.ext.autodoc',
+    'sphinx.ext.doctest',
+    'sphinx.ext.coverage',
+    'sphinx.ext.viewcode']
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']
