+++ docs/index.rst
@@ -1,10 +1,24 @@
 .. gitkv documentation master file, created by
-   sphinx-quickstart on Wed Mar 15 11:49:51 2017.
+   sphinx-quickstart on Thu Mar 23 12:49:22 2017.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
 Welcome to gitkv's documentation!
 =================================
+
+.. automodule:: gitkv
+
+Class Repo
+==========
+
+.. autoclass:: Repo
+    :members:
+
+Class FileInRepo
+================
+
+.. autoclass:: FileInRepo
+    :members:
 
 .. toctree::
    :maxdepth: 2
