+++ tests/__init__.py
@@ -1,9 +1,14 @@
 import gitkv
+
+
 def test_class_Repo():
-	with gitkv.Repo() as repo :
-		repo.os.makedirs('toto')
-		res = repo.os.path.exists('toto')
-	return res
+    with gitkv.Repo() as repo:
+        repo.os.makedirs('toto')
+        res = repo.os.path.exists('toto')
+    return res
+
+
 if __name__ == "__main__":
-	import doctest
-	doctest.testmod()
+    import doctest
+
+    doctest.testmod()
